backend/requirements.txt
fastapi 
uvicorn 
langchain
langchain-huggingface 
langchain-community 
pypdf 
python-multipart
python-dotenv 
transformers 
torch
faiss-cpu
numpy 
safetensors 
sentence-transformers 
tokenizers